Add files via upload
# -Ecommerce-Dashboard
# Data Analyst Internship Project: PowerBI Dashboard

## Project Overview:
This project was completed during my internship at ExcelR as a Data Analyst. It focuses on analyzing e-commerce data to extract insights into customer behavior and payment statistics. The dataset provided includes information about orders, payment types, review scores, delivery dates, and customer locations. The objective was to create a PowerBI dashboard to visualize these insights effectively.

## Dashboard Components:
### 1. Weekday Vs Weekend Payment Statistics:
Visualization comparing the number of orders with a review score of 5 and payment type as credit card between weekdays and weekends.

### 2. Average Delivery Time for Pet Shop Orders:
Analysis of the average number of days taken for orders to be delivered to customers for pet shop-related purchases.

### 3. Average Price and Payment Values in São Paulo City:
Insights into the average price and payment values from customers located in São Paulo city.

### 4. Shipping Days Vs Review Scores Relationship:
Exploration of the relationship between shipping days (order delivery time) and review scores, aiming to identify any correlation between them.

## Project Deliverables:
- PowerBI dashboard file (project.pbix)

## Instructions:
1. Download the PowerBI dashboard file (dashboard.pbix) from this GitHub repository.
2. Open the PowerBI dashboard file using PowerBI Desktop.
3. Refresh the dataset connection if necessary.
4. Navigate through the different tabs to explore the visualizations and insights.
5. Interact with the dashboard elements to filter and drill down into specific data points.
6. For any questions or feedback, please contact me at chakri.manepalli3@gmail.com.

## Note:
- Ensure that you have PowerBI Desktop installed to view and interact with the dashboard.

## Acknowledgements:
I would like to express my gratitude to ExcelR for providing me with the opportunity to work on this project during my internship.

## Disclaimer:
This dashboard and its insights are based on the provided dataset and are intended for educational and analytical purposes only. Any conclusions drawn should be carefully validated with additional data and domain knowledge.

## Author:
Akurathi Ajaykumar